+++ src/styles/themes/next-gen/variants/button.ts
@@ -4,6 +4,7 @@ import chroma from 'chroma-js';
 import { copyButton } from '../codeView/codeView';
 import tShirtSizes from '../customProperties/tShirtSizes';
 
+import { arrayFieldDeleteButton } from './arrayFieldDeleteButton';
 import { commonContentProps } from './box';
 
 
@@ -319,6 +320,9 @@ const primaryWithIcon = {
   ...primary,
   display: 'inline-flex',
   color: 'white',
+  '> span': {
+    color: 'white',
+  },
 };
 
 const inlineWithIcon = {
@@ -529,6 +533,9 @@ const iconButtons = {
   },
   onyx: {
     ...onyxIconButton,
+  },
+  arrayField: {
+    deleteButton: arrayFieldDeleteButton.deleteButton,
   },
   responseToolbar: {
     ...baseIconButton,