+++ src/styles/themes/next-gen/variants/variants.ts
@@ -2,6 +2,7 @@ import { astroTokens } from '@pingux/onyx-tokens';
 
 import attachment from '../../../../components/AIComponents/Attachment/Attachment.styles';
 import checkboxFieldGroup from '../../../../components/CheckboxFieldGroup/CheckboxFieldGroup.styles';
+import emptyState from '../../../../components/EmptyState/EmptyState.styles';
 import labelValuePairs from '../../../../components/LabelValuePairs/LabelValuePairs.styles';
 import skeleton from '../../../../components/Skeleton/Skeleton.styles';
 import codeView from '../codeView/codeView';
@@ -9,6 +10,7 @@ import sizes from '../sizes';
 
 import accordion from './accordion';
 import accordionGrid from './accordionGrid';
+import arrayField from './arrayField';
 import { avatar } from './avatar';
 import { box } from './box';
 import button, { defaultFocus } from './button';
@@ -32,7 +34,7 @@ import suggestion from './suggestion';
 import suggestions from './suggestions';
 import { table } from './table';
 import { tableBase } from './tableBase';
-import { menuTab, tab, tabs } from './tabs';
+import { menuTab, tab, tabPickerExpandIcon, tabs } from './tabs';
 import tooltip from './tooltip';
 
 const onyxLabelValuePairs = {
@@ -162,8 +164,8 @@ const listBox = {
     p: astroTokens.spacing.sm,
   },
   option: {
-    py: astroTokens.spacing['md-s'],
-    pl: astroTokens.spacing['md-s'],
+    py: 'md-s',
+    pl: 'md-s',
     pr: astroTokens.spacing.md,
     justifyContent: 'space-between',
     borderRadius: astroTokens.radius.md,
@@ -488,6 +490,7 @@ const helpHint = {
 export default {
   accordion,
   accordionGrid,
+  arrayField,
   attachment,
   avatar,
   breadcrumb,
@@ -497,6 +500,7 @@ export default {
   checkboxFieldGroup,
   codeView,
   dataTable,
+  emptyState,
   environmentBreadcrumb,
   fieldHelperText,
   footer,
@@ -512,6 +516,7 @@ export default {
   menuItem,
   menuSection,
   menuTab,
+  tabPickerExpandIcon,
   message,
   modal,
   navBar,