+++ src/styles/variants/variants.js
@@ -1,5 +1,6 @@
 import accordionGrid from '../../components/AccordionGridGroup/AccordionGrid.styles';
 import accordion from '../../components/AccordionGroup/Accordion.styles';
+import arrayField from '../../components/ArrayField/ArrayField.styles';
 import box from '../../components/Box/Box.styles';
 import bracket from '../../components/Bracket/Bracket.styles';
 import breadcrumb from '../../components/Breadcrumbs/Breadcrumb.styles';
@@ -12,6 +13,7 @@ import codeView from '../../components/CodeView/CodeView.styles';
 import collapsiblePanel from '../../components/CollapsiblePanel/CollapsiblePanel.styles';
 import copyText from '../../components/CopyText/CopyText.styles';
 import dataTable from '../../components/DataTable/DataTable.styles';
+import emptyState from '../../components/EmptyState/EmptyState.styles';
 import environmentBreadcrumb from '../../components/EnvironmentBreadcrumb/EnvironmentalBreadcrumb.styles';
 import expandableText from '../../components/ExpandableText/ExpandableText.styles';
 import fieldHelperText from '../../components/FieldHelperText/FieldHelperText.styles';
@@ -58,6 +60,7 @@ import treeView from '../../components/TreeView/TreeView.styles';
 export default {
   accordion,
   accordionGrid,
+  arrayField,
   box,
   bracket,
   breadcrumb,
@@ -71,6 +74,7 @@ export default {
   collapsiblePanel,
   copyText,
   dataTable,
+  emptyState,
   environmentBreadcrumb,
   expandableText,
   fieldHelperText,